requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
defusedxml==0.7.1
fonttools==4.42.0
fpdf==1.7.2
fpdf2==2.7.0
mysql-connector-python==8.1.0
numpy==1.25.2
Pillow==10.0.0
protobuf==4.21.12
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2023.3
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
